Added authentication to all pages and implemented likes and comments.

+++ handlers/comments.py
@@ -20,28 +20,26 @@ class AddComment(Handler):
 
     def post(self, post_id):
         post_id_int = int(post_id.split('/')[0])
-        user = User.by_id(int(self.get_cookie('user_id')))
-        if not user:
+        user_cookie = self.get_cookie('user_id')
+        if not user_cookie:
             errors = {
-                "unauth_error": "Please login to be able to comment."
+                "unauth_error": "Login error or Session timeout"
             }
             self.render('error.html', errors=errors)
         else:
-            post = Blog.get_by_id(post_id_int)
-            comment = self.request.get('comment')
-            comment = Comment(
-                post_id=post_id_int, username=user.name, comment=comment)
-            comment.put()
-            # Comments counter
-            # Default value is None
-            # If post has not comments, set it to one
-            # if post.comments is None:
-            #     post.comments = 1
-            # else:
-            #     post.comments = int(post.comments) + 1
-            # Update comments count
-            # post.put()
-            self.redirect('/blog/' + post_id)
+            user = User.by_id(int(user_cookie))
+            if not user:
+                errors = {
+                    "unauth_error": "Please login to be able to comment."
+                }
+                self.render('error.html', errors=errors)
+            else:
+                post = Blog.get_by_id(post_id_int)
+                comment = self.request.get('comment')
+                comment = Comment(
+                    post_id=post_id_int, username=user.name, comment=comment)
+                comment.put()
+                self.redirect('/blog/' + post_id)
 
 
 class DeleteComment(Handler):
@@ -55,19 +53,32 @@ class DeleteComment(Handler):
         """
         Delete the comment if the user is the commenter
         """
-        user = User.by_id(int(self.get_cookie('user_id')))
-        comment = Comment.get_by_id(int(comment_id))
-        if not user:
+        user_cookie = self.get_cookie('user_id')
+        if not user_cookie:
             errors = {
-                "unauth_error": "Please login to be able to comment."
+                "unauth_error": "Login error or Session timeout"
             }
             self.render('error.html', errors=errors)
         else:
-            if comment.username == user.name:
-                comment.delete()
-                self.redirect('/blog/' + post_id)
+            user = User.by_id(int(user_cookie))
+            if not user:
+                errors = {
+                    "unauth_error": "Please login to delete comments"
+                }
+                self.render('error.html', errors=errors)
             else:
-                self.redirect('/comment/error')
+                comment = Comment.get_by_id(int(comment_id))
+                if not comment:
+                    errors = {
+                        "unauth_error": "Comment does not exist"
+                    }
+                    self.render('error.html', errors=errors)
+                else:
+                    if comment.username == user.name:
+                        comment.delete()
+                        self.redirect('/blog/' + str(comment.post_id))
+                    else:
+                        self.redirect('/blog/comment/error')
 
 
 # TODO: Fix the EditComment functionality
@@ -78,28 +89,67 @@ class EditComment(Handler):
     post(): Handles editing of a comment on the blog handler
     """
 
-    def post(self, post_id):
-        post_id = post_id.split('/')[0]
-        key = db.Key.from_path('Blog', int(post_id), parent=blog_key())
-        post = db.get(key)
-        if post:
-            commentId = self.request.get('commentId')
-            editComment = self.request.get('editComment')
-            if commentId and editComment and self.user:
-                key = db.Key.from_path('Comment', int(
-                    commentId), parent=blog_key())
-                comment = db.get(key)
-                if comment:
-                    if comment.username == self.user.name:
-                        comment.comment = editComment
-                        comment.put()
-                        return self.redirect('/blog/' + post_id)
-                else:
-                    return self.redirect('/blog/' + post_id)
-            else:
-                return self.redirect('/')
+    def get(self, comment_id):
+        """
+        Shows edit comment page
+        """
+        user_cookie = self.get_cookie('user_id')
+        if not user_cookie:
+            errors = {
+                "unauth_error": "Login error or Session timeout"
+            }
+            self.render('error.html', errors=errors)
         else:
-            return self.redirect('/')
+            user = User.by_id(int(user_cookie))
+            if not user:
+                errors = {
+                    "unauth_error": "Please login to modify comments"
+                }
+                self.render('error.html', errors=errors)
+            else:
+                comment = Comment.get_by_id(int(comment_id))
+                if not comment:
+                    errors = {
+                        "unauth_error": "Comment not found!"
+                    }
+                    self.render('error.html', errors=errors)
+                else:
+                    if comment.username == user.name:
+                        self.render("editcomment.html", comment=comment)
+                    else:
+                        self.redirect('/blog/comment/error')
+
+    def post(self, comment_id):
+        """
+        Actual editing of the comment with the new content.
+        """
+        user_cookie = self.get_cookie('user_id')
+        if not user_cookie:
+            errors = {
+                "unauth_error": "Login error or Session timeout"
+            }
+            self.render('error.html', errors=errors)
+        else:
+            user = User.by_id(int(user_cookie))
+            if not user:
+                errors = {
+                    "unauth_error": "Please login to modify comments"
+                }
+                self.render('error.html', errors=errors)
+            else:
+                comment = Comment.get_by_id(int(comment_id))
+                if not comment:
+                    errors = {
+                        "unauth_error": "Comment not found!"
+                    }
+                    self.render('error.html', errors=errors)
+                else:
+                    if comment.username == user.name:
+                        comment.comment = self.request.get('comment')
+                        comment.put()
+                        self.redirect('/blog/' + str(comment.post_id))
+                    else:
+                        self.redirect('/blog/comment/error')
 
 
 class CommentError(Handler):
@@ -109,6 +159,6 @@ class CommentError(Handler):
         Handles error cases when unauthorized like is requested
         """
         errors = {
-            "unauth_error": "You can't comment your own post & can only like a post once."
+            "unauth_error": "Illegal access: Cannot modify other's comments."
         }
         self.render('error.html', errors=errors)
